refactor(store): extract format modules to files
