+++ rcommunity_core/src/store/format/tests.rs
@@ -1,0 +1,2 @@
+#[test]
+fn test_key_serialization() {}
